Add id = "player" for compatibility.
I'm using an extension called "Comments Sidebar for Youtube" and FastStream is broken whenever that extension is active. Adding id = "player" makes the player snap to the place it's supposed to, although I wish I could figure out how to make FastStream properly resize when the split bar is moved, at least going to fullscreen and back does the job at resizing it.